modulerized Options.Commander
