+++ app/Task/CLI.hs
@@ -17,7 +17,7 @@ import Control.Monad
 type TaskManager
   = Named "task-manager"
   & Annotated "the directory in which we keep our task files" (Env 'Optional "TASK_DIRECTORY" FilePath)
-    & ("help"
+    & (Sub "help"
       & Description "Displays this help text."
       & Raw
      + TaskProgram "edit" "Edits an already existing task. Fails if the task does not exist."
@@ -28,9 +28,9 @@ type TaskManager
      + Raw
     )
 
-type TaskProgram x desc = x & Description desc & Annotated ("the task we're going to " `AppendSymbol` x) (Arg "task-name" String) & Raw
+type TaskProgram x desc = Sub x & Description desc & Annotated ("the task we're going to " `AppendSymbol` x) (Arg "task-name" String) & Raw
 
-type TasklessProgram x desc = x & Description desc & Raw
+type TasklessProgram x desc = Sub x & Description desc & Raw
   
 taskManager :: ProgramT TaskManager IO ()
 taskManager = named @"task-manager" . annotated . envOptDef @"TASK_DIRECTORY" "tasks" $ \tasksFilePath ->
