+++ src/Options/Commander/InitialState.hs
@@ -1,0 +1,19 @@
+module Options.Commander.InitialState where
+
+
+-- | A simple default for getting out the arguments, options, and flags
+-- using 'getArgs'. We use the syntax ~flag for flags and -opt
+-- for options, with arguments using the typical ordered representation.
+initialState :: IO State
+initialState = do
+  args <- getArgs
+  let (opts, args', flags) = takeOptions args
+  return $ State args' (HashMap.fromList opts) (HashSet.fromList flags) 
+    where
+      takeOptions :: [String] -> ([(Text, Text)], [Text], [Text])
+      takeOptions = go [] [] [] where
+        go opts args flags (('~':x') : z) = go opts args (pack x' : flags) z
+        go opts args flags (('-':x) : y : z) = go ((pack x, pack y) : opts) args flags z
+        go opts args flags (x : y) = go opts (pack x : args) flags y
+        go opts args flags [] = (opts, reverse args, flags)
+
